the glorious spi_24L01 re-commit w linux non-sudo make
