+++ examples/spi_24L01_rx/Makefile
@@ -1,11 +1,11 @@
+all : flash
+
 TARGET:=spi_24L01_rx
-ADDITIONAL_C_FILES+=SPI.c nrf24l01_low_level.c nrf24l01.c 
+ADDITIONAL_C_FILES+=nrf24l01_low_level.c nrf24l01.c 
 
 CFLAGS+=-DSTDOUT_UART
 
 include ../../ch32v003fun/ch32v003fun.mk
 
-all : flash
 flash : cv_flash
 clean : cv_clean
-
