+++ examples/spi_24L01_rx/spi_24L01_rx.c
@@ -3,8 +3,6 @@
  * 04-26-2023 recallmenot 
  */
 
-// Could be defined here, or in the processor defines.
-#include <stdint.h>
 #define SYSTEM_CORE_CLOCK 48000000
 #define APB_CLOCK SYSTEM_CORE_CLOCK
 
@@ -12,14 +10,47 @@
 #include <stdio.h>
 #include "nrf24l01.h"
 
-#define TIME_GAP			300
 
+
+#define TIME_GAP 300
 uint8_t ascending_number = 0;
 char txt[16];
 
 
 
-// ####### start of main program ###############################################################
+//######### debug fn
+
+void uint8_to_binary_string(uint8_t value, char* output, int len) {
+		for (int i = 0; i < len; i++) {
+				output[len - i - 1] = (value & 1) ? '1' : '0';
+				value >>= 1;
+		}
+		output[len] = '\0';
+}
+
+
+void print_reg(char* name, uint8_t addr) {
+	char str[9];
+	uint8_t REG;
+	nrf24_read(addr, &REG, 1, CLOSE);
+	uint8_to_binary_string(REG, str, 8);
+	printf("				 %s register: %s\n\r", name, str);
+}
+
+
+void print_debug() {
+	print_reg("FEATURE      ", FEATURE_ADDRESS);
+	print_reg("TX OBSERVE   ", OBSERVE_TX_ADDRESS);
+	print_reg("STATUS       ", STATUS_ADDRESS);
+	print_reg("RX_PW_P0 ADDR", RX_ADDR_P0_ADDRESS);
+	print_reg("TX ADDR      ", TX_ADDR_ADDRESS);
+	print_reg("EN_AA        ", EN_AA_ADDRESS);
+	print_reg("EN_RXADDR    ", EN_RXADDR_ADDRESS);
+}
+
+
+
+//######### LED fn
 
 // wire PD4 to LED1 on the dev board (-)
 void led_on() {
@@ -31,6 +62,8 @@ void led_off() {
 }
 
 
+
+//######### RX fn
 
 uint8_t recvnumber() {
   return nrf24_receive(&ascending_number, 1);
@@ -58,13 +91,17 @@ void receive() {
 		case OPERATION_DONE:
 			led_on();
 			// pick one of these two:
-			printf("***   RX success, received: %u\n\r", ascending_number);
-			//printf("***   RX success, received: %s\n\r", txt);
+			//printf("***   RX success, received: %u\n\r", ascending_number);
+			printf("***   RX success, received: %s\n\r", txt);
 			led_off();
 			break;
 	}
 	Delay_Ms(TIME_GAP);
 }
+
+
+
+//######### MAIN
 
 int main()
 {
@@ -74,16 +111,21 @@ int main()
 	SetupUART( UART_BRR );
 	Delay_Ms( 100 );
 
+	printf("\r\r\n\nspi_24L01_RX\n\r");
+
+	printf("initializing radio as RX...");
+	nrf24_device(RECEIVER, RESET);
+	nrf24_rf_power(18);						//default TX power is -6dB, pretty strong, reduce to -18dBm for one room (ACK = TX)
+	printf("done.\n\r");
+
+	print_debug();
+
 	// GPIO D0 Push-Pull for RX notification
 	RCC->APB2PCENR |= RCC_APB2Periph_GPIOD;
 	GPIOD->CFGLR &= ~(0xf<<(4*4));
 	GPIOD->CFGLR |= (GPIO_Speed_10MHz | GPIO_CNF_OUT_PP)<<(4*4);
 
-	printf("\r\r\n\nspi_24L01_RX\n\r");
-
-	printf("initializing radio as RX...");
-	nrf24_device(RECEIVER, RESET);
-	printf("done.\n\r");
+	Delay_Ms(1000);
 
 	printf("looping...\n\r");
 	while(1)
